requirements-api.txt
fastapi[standard]>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
